+++ src/emu/nosv/setup.c
@@ -57,6 +57,8 @@ static struct ev_decl model_evlist[] = {
 	PAIR_E("VAy", "VAY", "nosv_yield()")
 	PAIR_E("VAw", "VAW", "nosv_waitfor()")
 	PAIR_E("VAc", "VAC", "nosv_schedpoint()")
+	PAIR_E("VAa", "VAA", "nosv_attach()")
+	PAIR_E("VAe", "VAE", "nosv_detach()")
 
 	/* FIXME: VHA and VHa are not subsystems */
 	{ "VHa", "enters nosv_attach()" },
@@ -70,7 +72,7 @@ static struct ev_decl model_evlist[] = {
 
 struct model_spec model_nosv = {
 	.name    = model_name,
-	.version = "1.0.0",
+	.version = "1.1.0",
 	.evlist  = model_evlist,
 	.model   = model_id,
 	.create  = model_nosv_create,
@@ -135,7 +137,8 @@ static const struct pcf_value_label nosv_ss_values[] = {
 	{ ST_API_YIELD,        "API: Yield" },
 	{ ST_API_WAITFOR,      "API: Waitfor" },
 	{ ST_API_SCHEDPOINT,   "API: Scheduling point" },
-	{ ST_ATTACH,           "Thread: Attached" },
+	{ ST_API_ATTACH,       "API: Attach" },
+	{ ST_API_DETACH,       "API: Detach" },
 	{ ST_WORKER,           "Thread: Worker" },
 	{ ST_DELEGATE,         "Thread: Delegate" },
 	{ EV_SCHED_SEND,       "EV Scheduler: Send task" },